employeeSystem 정의

+++ src/main/java/com/nastech/upmureport/domain/entity/UserProjectPK.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 public class UserProjectPK implements Serializable {
 	private static final long serialVersionUID = -1975783332618600132L;
 	
-	private Long user;
+	private Integer user;
 	private Long project;
 	
 	@Override
